test(e2e): stabilize production journeys
